Create lex-pipe.ts
