+++ lib/syntax/lex-pipe.d.ts
@@ -1,0 +1,6 @@
+import { Frame, FrameString } from "../frames";
+export declare class LexPipe extends Frame {
+    constructor(out: Frame);
+    lex_string(input: string): Frame;
+    lex(source: FrameString): Frame;
+}
