+++ lib/syntax/lex-pipe.js
@@ -1,0 +1,26 @@
+"use strict";
+var __extends = (this && this.__extends) || function (d, b) {
+    for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p];
+    function __() { this.constructor = d; }
+    d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
+};
+var frames_1 = require("../frames");
+var tokens_1 = require("./tokens");
+var LexPipe = (function (_super) {
+    __extends(LexPipe, _super);
+    function LexPipe(out) {
+        var _this = this;
+        tokens_1.tokens[frames_1.Frame.kOUT] = out;
+        _this = _super.call(this, tokens_1.tokens) || this;
+        return _this;
+    }
+    LexPipe.prototype.lex_string = function (input) {
+        var source = new frames_1.FrameString(input);
+        return this.lex(source);
+    };
+    LexPipe.prototype.lex = function (source) {
+        return source.reduce(this);
+    };
+    return LexPipe;
+}(frames_1.Frame));
+exports.LexPipe = LexPipe;
