+++ lib/syntax/pipes.js
@@ -5,25 +5,7 @@ var __extends = (this && this.__extends) || function (d, b) {
     d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
 };
 var frames_1 = require("../frames");
-var tokens_1 = require("./tokens");
-var LexPipe = (function (_super) {
-    __extends(LexPipe, _super);
-    function LexPipe(out) {
-        var _this = this;
-        tokens_1.tokens[frames_1.Frame.kOUT] = out;
-        _this = _super.call(this, tokens_1.tokens) || this;
-        return _this;
-    }
-    LexPipe.prototype.lex_string = function (input) {
-        var source = new frames_1.FrameString(input);
-        return this.lex(source);
-    };
-    LexPipe.prototype.lex = function (source) {
-        return source.reduce(this);
-    };
-    return LexPipe;
-}(frames_1.Frame));
-exports.LexPipe = LexPipe;
+var lex_pipe_1 = require("./lex-pipe");
 var EvalPipe = (function (_super) {
     __extends(EvalPipe, _super);
     function EvalPipe(out, meta) {
@@ -53,7 +35,7 @@ var piper = function (input, context) {
     var result = new frames_1.FrameArray([], context);
     var evaluator = new EvalPipe(result);
     var parser = new ParsePipe(evaluator);
-    var lexer = new LexPipe(parser);
+    var lexer = new lex_pipe_1.LexPipe(parser);
     var status = lexer.lex_string(input);
     if (status !== lexer) {
     }
