+++ test/syntax/lex-pipe-spec.ts
@@ -1,5 +1,5 @@
 import { expect } from "chai";
-import { LexPipe } from "../../src/syntax/pipes";
+import { LexPipe } from "../../src/syntax/lex-pipe";
 
 describe("LexPipe", () => {
   it("is exported", () => {
